Move method from `AbstractTask` to `DefaultTask`
