+++ subprojects/core/src/main/java/org/gradle/api/internal/AbstractTask.java
@@ -25,14 +25,10 @@ import org.codehaus.groovy.runtime.InvokerInvocationException;
 import org.gradle.api.Action;
 import org.gradle.api.AntBuilder;
 import org.gradle.api.Describable;
-import org.gradle.api.Incubating;
 import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.file.DirectoryProperty;
-import org.gradle.api.file.RegularFileProperty;
 import org.gradle.api.internal.file.FileResolver;
-import org.gradle.api.internal.file.TaskFileVarFactory;
 import org.gradle.api.internal.file.TemporaryFileProvider;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.internal.tasks.ClassLoaderAwareTaskAction;
@@ -993,49 +989,5 @@ public abstract class AbstractTask implements TaskInternal, DynamicObjectAware {
     @Override
     public boolean isHasCustomActions() {
         return hasCustomActions;
-    }
-
-    /**
-     * Creates a new output directory property for this task.
-     *
-     * @return The property.
-     * @since 4.1
-     */
-    @Incubating
-    protected DirectoryProperty newOutputDirectory() {
-        return getServices().get(TaskFileVarFactory.class).newOutputDirectory(this);
-    }
-
-    /**
-     * Creates a new output file property for this task.
-     *
-     * @return The property.
-     * @since 4.1
-     */
-    @Incubating
-    protected RegularFileProperty newOutputFile() {
-        return getServices().get(TaskFileVarFactory.class).newOutputFile(this);
-    }
-
-    /**
-     * Creates a new input file property for this task.
-     *
-     * @return The property.
-     * @since 4.1
-     */
-    @Incubating
-    protected RegularFileProperty newInputFile() {
-        return getServices().get(TaskFileVarFactory.class).newInputFile(this);
-    }
-
-    /**
-     * Creates a new input directory property for this task.
-     *
-     * @return The property.
-     * @since 4.3
-     */
-    @Incubating
-    protected DirectoryProperty newInputDirectory() {
-        return getServices().get(TaskFileVarFactory.class).newInputDirectory(this);
     }
 }
